[FIX] [#30] Refactor confirmation token implementation
